+++ core/pom.xml
@@ -46,6 +46,14 @@
                 <groupId>org.eclipse.jetty</groupId>
                 <artifactId>jetty-maven-plugin</artifactId>
                 <version>9.2.11.v20150529</version>
+                <configuration>
+                    <webApp>
+                        <contextPath>/core</contextPath>
+                    </webApp>
+                    <httpConnector>
+                        <port>8083</port>
+                    </httpConnector>
+                </configuration>
             </plugin>
         </plugins>
     </build>
