+++ monitor/src/main/java/com/owera/xaps/monitor/Properties.java
@@ -1,10 +1,15 @@
 package com.owera.xaps.monitor;
 
 import java.util.Map;
+import java.util.Optional;
+import java.util.function.Supplier;
 
+import com.owera.common.db.ConnectionProperties;
 import com.owera.common.util.PropertyReader;
 import com.typesafe.config.Config;
 import com.typesafe.config.ConfigFactory;
+
+import static java.lang.Long.getLong;
 
 public class Properties {
 
@@ -72,5 +77,24 @@ public class Properties {
 		}
 		return prop;
 	}
+
+	public static int getMaxConn(final String infix) {
+		return getInteger("db." + infix + ".maxconn", ConnectionProperties.maxconn);
+	}
+
+	public static long getMaxAge(final String infix) {
+		return getLong("db." + infix + ".maxage", ConnectionProperties.maxage);
+	}
+
+	public static String getUrl(final String infix) {
+		return Optional.ofNullable(getString("db." + infix + ".url", null))
+				.orElseGet(new Supplier<String>() {
+					@Override
+					public String get() {
+						return getString("db." +infix, null);
+					}
+				});
+	}
+
 
 }
