+++ monitor/src/main/java/com/owera/xaps/monitor/task/TriggerNotificationHourly.java
@@ -17,6 +17,10 @@ import com.owera.xaps.dbi.User;
 import com.owera.xaps.dbi.Users;
 import com.owera.xaps.monitor.MonitorServlet;
 
+import static com.owera.xaps.monitor.Properties.getMaxAge;
+import static com.owera.xaps.monitor.Properties.getMaxConn;
+import static com.owera.xaps.monitor.Properties.getUrl;
+
 public class TriggerNotificationHourly extends TaskDefaultImpl {
 
 	private Logger log = new Logger();
@@ -45,7 +49,7 @@ public class TriggerNotificationHourly extends TaskDefaultImpl {
 		Users users = new Users(xapsCp);
 		User user = users.getUnprotected(Users.USER_ADMIN);
 		Identity id = new Identity(SyslogConstants.FACILITY_STUN, MonitorServlet.VERSION, user);
-		ConnectionProperties sysCp = ConnectionProvider.getConnectionProperties("xaps-monitor.properties", "db.syslog");
+		ConnectionProperties sysCp = ConnectionProvider.getConnectionProperties(getUrl("syslog"), getMaxAge("syslog"), getMaxConn("syslog"));
 		Syslog syslog = new Syslog(sysCp, id);
 		return new DBI(Integer.MAX_VALUE, xapsCp, syslog);
 	}
