+++ monitor/src/main/resources/xaps-monitor.conf
@@ -41,8 +41,8 @@ alerts.from		 	 = "from-email@address.com"
 # be sent if the whole server goes down.
 # Default is "http://localhost/"
 
-# monitor.url.core = http://anotherhost/core/ok
-# monitor.url.web  = http://yetanotherhost/web/ok
+monitor.url.core = "http://localhost:8083/core/ok"
+monitor.url.web  = "http://localhost:8081/web/ok"
 monitor.urlbase 	= "http://localhost/"
 
 # Email URL
