+++ web/src/main/resources/xaps-web.conf
@@ -16,7 +16,7 @@ session.timeout = 60
 
 # --- Monitor server ---
 # where to retrieve monitor status (not including context root and other parameters)
-monitor.location = "http://localhost/"
+monitor.location = "http://localhost:8082/monitor"
 
 # --- Database connections ---
 db {
